requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# bot.py: 3,6,7,8
# cogs\help.py: 1,2
# cogs\moderation.py: 1,2
# cogs\owner.py: 1,2
# cogs\ping.py: 1,4
discord.py
